+++ app/Http/Controllers/Web/BlogPostController.php
@@ -16,7 +16,13 @@ class BlogPostController extends Controller
     }
     public function create()
     {
-        return view('blog.create');
+        if (auth()->user()->hasPermissionTo('create-blog-posts')) {
+
+
+            return view('blog.create');
+        } else {
+            return redirect()->route('blog.index')->withErrors(['You do not have permission to create a blog']);
+        }
     }
 
     public function show(BlogPost $blog) {
@@ -31,9 +37,16 @@ class BlogPostController extends Controller
         return redirect()->route('blog.edit', $blog->id);
     }
 
-    public function edit(BlogPost $blog): \Illuminate\Contracts\View\View|\Illuminate\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\Foundation\Application
+    public function edit(BlogPost $blog)
     {
-        return view('blog.edit',compact('blog'));
+
+        $user = auth()->user();
+        if ($user->hasPermissionTo('edit-blog-posts') && $user === $blog->user()) {
+            return view('blog.edit', compact('blog'));
+        }
+        else {
+            return redirect()->route('blog.index')->withErrors(['You do not have access to edit the blog']);
+        }
     }
 
     public function update(BlogPostRequest $request, BlogPost $blog)
@@ -43,7 +56,15 @@ class BlogPostController extends Controller
     }
     public function destroy(BlogPost $blog)
     {
-        $blog->delete();
-        return redirect()->route('blog.index')->with('success','Company has been deleted successfully');
+        $user = auth()->user();
+        if ($user->hasPermissionTo('delete-blog-posts') && $user === $blog->user()) {
+
+            $blog->delete();
+            return redirect()->route('blog.index')->with('success','Company has been deleted successfully');
+
+        }
+        else {
+            return redirect()->route('blog.index')->withErrors(['You cannot delete the file']);
+        }
     }
 }
